System.TimeIt

Documentation

timeIt :: MonadIO m => m a -> m a

Wrap a MonadIO computation so that it prints out the execution time.

timeItShow :: (MonadIO m, Show a) => m a -> m a

Like timeIt, but uses the show rendering of a as label for the timing.

Since: 2.0

timeItNamed :: MonadIO m => String -> m a -> m a

Like timeIt, but uses the String as label for the timing.

Since: 2.0

timeItT :: MonadIO m => m a -> m (Double, a)

Wrap a MonadIO computation so that it returns execution time in seconds, as well as the result value.